views: Move API views into "api" package
