+++ skylines/api/views/notifications.py
@@ -4,7 +4,7 @@ from sqlalchemy.sql.expression import or_
 
 from skylines.api.json import jsonify
 from skylines.database import db
-from skylines.frontend.oauth import oauth
+from skylines.api.oauth import oauth
 from skylines.model import Event, Notification, Flight, User
 
 TYPES = {
