+++ skylines/api/views/ranking.py
