+++ skylines/api/views/statistics.py
