--- skylines/api/views/user.py
@@ -1,16 +1,0 @@
-from flask import Blueprint, request
-
-from skylines.model import User
-from skylines.api.oauth import oauth
-from skylines.api.schemas import current_user_schema
-from skylines.api.json import jsonify
-
-user = Blueprint('user', 'skylines')
-
-
-@user.route('/user', strict_slashes=False)
-@oauth.required()
-def read():
-    user = User.get(request.user_id)
-    result = current_user_schema.dump(user)
-    return jsonify(result.data)
